Improve ux in step sliders (#26)

+++ app/lib/model/led_effects.dart
@@ -97,7 +97,7 @@ class BreathingLedEffect extends LedEffect implements StorableObject {
   EffectType get type => EffectType.breathing;
 
   HSVColor color;
-  double step;
+  int step;
   double peak;
   bool breatheFromOff;
